refactor: extract AddCreditsModal into separate component file
